requirements.txt
streamlit==1.31.0
langchain==0.0.267
langchain-openai==0.0.2
python-dotenv==1.0.0
openai==1.3.0